Moved SoundEffect into a separate compilation unit.
Also, fixed license header in TabletScriptingInterface, to include proper author and date.
